+++ NEWS.md
@@ -1,5 +1,8 @@
 # teal 0.15.2.9018
 
+### Miscellaneous
+* Filter mapping display is no longer coupled to the snapshot manager.
+
 # teal 0.15.2
 
 ### Bug fixes